Adding two interactive functions to change rubies
`chruby-use` takes a string and will try to activate it, if it can't
find anything to activate it will display a message to the user.

`chruby-user-corresponding` will look for a `.ruby-version` file in the
directory tree and tries to activate that ruby.

Both are inspired by `rbenv.el`'s functions and differ only slightly.

Update: replaced if with when.